LayoutTests/css3/filters/custom-with-at-rule-syntax/parsing-mix-property-invalid-expected.txt

Test mix property parsing in the @-webkit-filter at-rule.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



========================================
Value test.
========================================

No property value.
mix: ;
PASS filterAtRule.type is CSSRule.WEBKIT_FILTER_RULE
PASS styleDeclaration.length is 0
PASS styleDeclaration.getPropertyValue('mix') is null

Three values.
mix: screen source-over screen;
PASS filterAtRule.type is CSSRule.WEBKIT_FILTER_RULE
PASS styleDeclaration.length is 0
PASS styleDeclaration.getPropertyValue('mix') is null

Three values.
mix: screen source-over source-over;
PASS filterAtRule.type is CSSRule.WEBKIT_FILTER_RULE
PASS styleDeclaration.length is 0
PASS styleDeclaration.getPropertyValue('mix') is null

Two blend modes.
mix: screen screen;
PASS filterAtRule.type is CSSRule.WEBKIT_FILTER_RULE
PASS styleDeclaration.length is 0
PASS styleDeclaration.getPropertyValue('mix') is null

Two composite operators.
mix: source-over source-over;
PASS filterAtRule.type is CSSRule.WEBKIT_FILTER_RULE
PASS styleDeclaration.length is 0
PASS styleDeclaration.getPropertyValue('mix') is null

Two blend modes.
mix: screen screen;
PASS filterAtRule.type is CSSRule.WEBKIT_FILTER_RULE
PASS styleDeclaration.length is 0
PASS styleDeclaration.getPropertyValue('mix') is null

Comma between values.
mix: screen, source-over;
PASS filterAtRule.type is CSSRule.WEBKIT_FILTER_RULE
PASS styleDeclaration.length is 0
PASS styleDeclaration.getPropertyValue('mix') is null
PASS successfullyParsed is true

TEST COMPLETE
